More work on making buttons work correctly.
The screen can now scroll down farther such that no part of the map is
inaccessable  due to the HUD. The mouse can no longer drag around or through the
HUD.

Button still won't make a unit appear; further work is needed in that regard.

Added sprites for spawn heavy tank (tank destroyer?) and artillery (howitzer)
as well as colored sprites for the arty.